move javascriptForUI in different folder
